Update Maven publishing workflow in GitHub Actions
The GitHub Actions configuration for Maven publishing has been updated. The publishing job was renamed from 'build' to 'publish', and the action used was changed from a specific commit-hash to version 2 of the 'gradle-build-action'. Alongside these, Gradle version was revised to 8.5 and finally, an explicit step for running 'publishAllPublicationsToSonatypeRepository' was added.